% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapscape.R
\name{mapscapeOutput}
\alias{mapscapeOutput}
\title{Widget output function for use in Shiny}
\usage{
mapscapeOutput(outputId, width = "100\%", height = "400px")
}
\description{
Widget output function for use in Shiny
}

