% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapscape.R
\name{renderMapscape}
\alias{renderMapscape}
\title{Widget render function for use in Shiny}
\usage{
renderMapscape(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}

