% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacesweep.R
\name{renderSpacesweep}
\alias{renderSpacesweep}
\title{Widget render function for use in Shiny}
\usage{
renderSpacesweep(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}

