% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacesweep.R
\name{spacesweepOutput}
\alias{spacesweepOutput}
\title{Widget output function for use in Shiny}
\usage{
spacesweepOutput(outputId, width = "100\%", height = "400px")
}
\description{
Widget output function for use in Shiny
}

