% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapscape.R
\name{mapscapeOutput}
\alias{mapscapeOutput}
\title{Widget output function for use in Shiny}
\usage{
mapscapeOutput(outputId, width = "100\%", height = "400px")
}
\arguments{
\item{outputId}{-- id of output}

\item{width}{-- width of output}

\item{height}{-- height of output}
}
\description{
Widget output function for use in Shiny
}
\examples{
mapscapeOutput(1, '100\%', '300px')
mapscapeOutput(1, '80\%', '300px')
}

