% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapscape.R
\name{renderMapscape}
\alias{renderMapscape}
\title{Widget render function for use in Shiny}
\usage{
renderMapscape(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{-- expression for Shiny}

\item{env}{-- environment for Shiny}

\item{quoted}{-- default is FALSE}
}
\description{
Widget render function for use in Shiny
}
\examples{
mapscapeOutput("hi")
mapscapeOutput("hello")
}

