% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacesweep.R
\name{spacesweep}
\alias{spacesweep}
\title{SpaceSweep}
\usage{
spacesweep(clonal_prev, tree_edges, clone_colours = "NA", mutations = "NA",
  mutation_prevalences = "NA", gender, site_ids = "NA", show_root = FALSE,
  n_cells = 100, width = 960, height = 960)
}
\arguments{
\item{clonal_prev}{Clonal prevalence data frame.
Format: columns are (1) {String} "site_id" - id for the anatomic site
                    (2) {String} "clone_id" - clone id
                    (3) {Number} "clonal_prev" - clonal prevalence.}

\item{tree_edges}{Tree edges data frame. The root of the tree (id: "Root") must be specified as a source.
Format: columns are (1) {String} "source" - source node id
                    (2) {String} "target" - target node id.}

\item{clone_colours}{(Optional) Data frame with clone ids and their corresponding colours 
Format: columns are (1) {String} "clone_id" - the clone ids
                    (2) {String} "colour" - the corresponding Hex colour for each clone id.}

\item{mutations}{(Optional) Data frame of mutations occurring before the appearance of each clone.
Format: columns are (1) {String} "chrom" - chromosome number
                    (2) {Number} "coord" - coordinate of mutation on chromosome
                    (3) {String} "clone_id" - clone id
                    (4) {String} "gene_name" - name of the affected gene (can be "" if none affected).}

\item{mutation_prevalences}{(Optional) Data frame of mutation prevalences for each anatomic site in a patient.
Format: columns are (1) {String} "chrom" - chromosome number
                    (2) {Number} "coord" - coordinate of mutation on chromosome
                    (3) {String} "site_id" - anatomic site id
                    (4) {Number} "prev" - prevalence of the mutation
                    (5) {String} (Optional) "effect" - effect of the mutation 
                                                       (e.g. non-synonymous, upstream, etc.)
                    (5) {String} (Optional) "impact" - impact of the mutation (e.g. low, moderate, high).}

\item{gender}{Gender of the patient (M/F).}

\item{site_ids}{{Vector} (Optional) Ids of the sites in the order your wish to display them 
(clockwise from positive x-axis).}

\item{show_root}{(Optional) Whether or not to plot the root (for tree).}

\item{n_cells}{(Optional) The number of cells to plot (for voronoi tessellation).}
}
\description{
Generates patient clonal SpaceSweeps.
}

