% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacesweep.R
\name{spacesweep}
\alias{spacesweep}
\title{SpaceSweep}
\usage{
spacesweep(clonal_prev, tree_edges, clone_colours, show_root = TRUE,
  n_cells = 100, width = 960, height = 960)
}
\arguments{
\item{clonal_prev}{Clonal prevalence data frame.
Format: columns are (1) {String} "site_id" - id for the anatomic site
                    (2) {String} "clone_id" - clone id
                    (3) {Number} "clonal_prev" - clonal prevalence.}

\item{tree_edges}{Tree edges data frame. The root of the tree (id: "Root") must be specified as a source.
Format: columns are (1) {String} "source" - source node id
                    (2) {String} "target" - target node id.}

\item{clone_colours}{Data frame with clone ids and their corresponding colours 
Format: columns are (1) {String} "clone_id" - the clone ids
                    (2) {String} "colour" - the corresponding Hex colour for each clone id.}

\item{show_root}{Whether or not to plot the root (for tree).}

\item{n_cells}{The number of cells to plot (for voronoi tessellation).}
}
\description{
Generates patient clonal SpaceSweeps.
}

